fix: AvailDAService new func for new avail rust sdk

+++ crates/adapters/avail/src/types/config.rs
@@ -4,7 +4,7 @@ use schemars::JsonSchema;
 pub struct AvailDAConfig {
     pub http_api_url: String,
     pub ws_api_url: String,
-    pub proof_app_id: u64,
-    pub batch_app_id: u64,
+    pub proof_app_id: u32,
+    pub batch_app_id: u32,
     pub signer_key: String,
 }
